FIX Uses application.conf to inject version
The preivous approach to set the version at build time and read it at
compilation time wasn't working. So now we defined the version once at
`application.conf`, and we parse this file to get the version and
inject it back into `build.sbt`.

The version is defined on a single place and readable at RunTime
without reflection.